[REFACTOR]: refactor
