--- click_me_to_run.bat
@@ -1,2 +1,0 @@
-conda activate angle-detector
-python infer.py
