--- cpp/config.h
@@ -1,2 +1,0 @@
-#pragma once
-#define pi acos(-1)
