--- requirements.txt
@@ -1,51 +1,0 @@
-altgraph==0.17.4
-certifi==2024.8.30
-charset-normalizer==3.4.0
-colorama==0.4.6
-coloredlogs==15.0.1
-contourpy==1.3.1
-cycler==0.12.1
-filelock==3.16.1
-flatbuffers==24.3.25
-fonttools==4.55.2
-fsspec==2024.10.0
-humanfriendly==10.0
-idna==3.10
-Jinja2==3.1.4
-kiwisolver==1.4.7
-MarkupSafe==3.0.2
-matplotlib==3.9.3
-mpmath==1.3.0
-networkx==3.4.2
-numpy==2.2.0
-onnx==1.17.0
-onnxruntime==1.20.1
-opencv-python==4.10.0.84
-packaging==24.2
-pandas==2.2.3
-pefile==2023.2.7
-Pillow==9.5.0
-protobuf==5.29.1
-psutil==6.1.0
-py-cpuinfo==9.0.0
-pyinstaller==6.11.1
-pyinstaller-hooks-contrib==2024.10
-pyparsing==3.2.0
-pyreadline3==3.5.4
-python-dateutil==2.9.0.post0
-pytz==2024.2
-pywin32-ctypes==0.2.3
-PyYAML==6.0.2
-requests==2.32.3
-scipy==1.14.1
-seaborn==0.13.2
-six==1.17.0
-sympy==1.13.1
-torch==2.5.1
-torchvision==0.20.1
-tqdm==4.67.1
-typing_extensions==4.12.2
-tzdata==2024.2
-ultralytics==8.3.48
-ultralytics-thop==2.0.13
-urllib3==2.2.3
