--- res/data.txt
@@ -1,5 +1,0 @@
-CenterPoint 542.7611075,921.5313699999999
-Angle 0.818444508088514
-Position above
-Diameter 702.0399772485654
-SliderAngle 1.3369034892860725
